+++ lib/ratehawk/index.ts
@@ -2,4 +2,9 @@ export * as hotel from "./services/hotel";
 export { ratehawkRequest, RatehawkError } from "./client";
 export { config as ratehawkConfig } from "./config";
 export { rhImage } from "./image";
+export {
+  getCachedHotelInfo,
+  setCachedHotelInfo,
+  hotelInfoBatchCached,
+} from "./cache";
 export type * from "./types";